+++ CHANGES.rst
@@ -12,6 +12,7 @@ Added
 - Configuration entries can be loaded from files if the entry key has a *_FILE* suffix
   and the entry value is the path to the file. :issue:`134` :pr:`134`
 - Field list support. :issue:`115` :pr:`136`
+- Pages are boosted with HTMX :issue:`144` :issue:`145` :pr:`137`
 
 Removed
 *******